refactor setup for magnification_finite_adaptive
